+++ CHANGELOG.md
@@ -4,6 +4,10 @@
   * [FIX]: 'Create MiniReel' modals DOM update: 'General' tab added first, 'AutoPlay Settings' added to the end.
   * Ad server tab is only shown if the org is provisioned with their own
     ad server
+  * Ad cards are now auto-added when creating a new MiniReel to meet the
+    org's minAdCount
+  * Ad cards will not be deleteable when the org's minAdCount is
+    reached
 
 ## Beta5 (May 20, 2014)
 * *[Beta5.rc1]*